backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
attrs==25.4.0
Authlib==1.6.5
autobahn==25.10.2
Automat==25.4.16
banks==2.2.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
channels==4.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.3
daphne==4.0.0
dataclasses-json==0.6.7
Deprecated==1.3.1
deprecation==2.1.0
dirtyjson==1.0.8
distro==1.9.0
Django==5.0.6
django-cors-headers==4.8.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
filetype==1.2.0
frozenlist==1.8.0
fsspec==2025.10.0
google-auth==2.43.0
greenlet==3.2.4
griffe==1.14.0
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
inflection==0.5.1
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
llama-index-core==0.14.8
llama-index-embeddings-ollama==0.8.4
llama-index-instrumentation==0.4.2
llama-index-llms-ollama==0.9.0
llama-index-vector-stores-weaviate==1.4.1
llama-index-workflows==2.10.3
llamaindex-py-client==0.1.19
MarkupSafe==3.0.3
marshmallow==3.26.1
multidict==6.7.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.2
numpy==2.3.4
ollama==0.6.0
openai==2.7.2
packaging==25.0
pandas==2.3.3
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
propcache==0.4.1
protobuf==6.33.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
pytest==9.0.0
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlparse==0.5.3
tenacity==8.5.0
tiktoken==0.12.0
tqdm==4.67.1
Twisted==25.5.0
txaio==25.9.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
validators==0.35.0
weaviate-client==4.18.0
wrapt==2.0.0
yarl==1.22.0
zope.interface==8.1
llama-index-core
llama-index-vector-stores-weaviate
llama-index-llms-ollama
llama-index-embeddings-ollama
ollama
weaviate-client
httpx